+++ packages/api/src/lib/inferenceNodeApi/infer.ts
@@ -1,7 +1,6 @@
 import { env } from "@/env.js";
-import { toMilliseconds } from "duration-fns";
 import { FetchError, ResponseOkError } from "../errors.js";
-import { abortSignal, filterWhitespaceStrings } from "../utils.js";
+import { filterWhitespaceStrings } from "../utils.js";
 import { v } from "../valibot.js";
 
 export const OptionsSchema = v.object({
@@ -40,6 +39,8 @@ const InferenceSchema = v.object({
 const EpilogueSchema = v.object({
   type: v.literal("Epilogue"),
   duration: v.number(),
+  aborted: v.boolean(),
+  tokenLength: v.number(),
   contextLength: v.number(),
 });
 
@@ -53,9 +54,7 @@ export async function* infer(
   baseUrl: string,
   sessionId: string,
   args: v.InferInput<typeof RequestBodySchema>,
-  options: { timeout: number } = {
-    timeout: toMilliseconds({ minutes: 2 }),
-  },
+  options?: { abortSignal: AbortSignal },
 ): AsyncGenerator<v.InferOutput<typeof ChunkSchema>> {
   let response;
   try {
@@ -66,7 +65,7 @@ export async function* infer(
         "Content-Type": "application/json",
       },
       body: JSON.stringify({ ...args, stream: true }),
-      signal: abortSignal(options.timeout),
+      signal: options?.abortSignal,
     });
   } catch (e: any) {
     throw new FetchError(e.message);