+++ packages/api/src/web/routes/gpts/create.ts
@@ -3,12 +3,13 @@ import { d } from "@/lib/drizzle.js";
 import { FetchError, ResponseOkError } from "@/lib/errors.js";
 import * as inferenceNodeApi from "@/lib/inferenceNodeApi.js";
 import { konsole } from "@/lib/konsole.js";
-import { unreachable } from "@/lib/utils.js";
+import { timeoutSignal, unreachable } from "@/lib/utils.js";
 import { v } from "@/lib/valibot.js";
 import bodyParser from "body-parser";
 import cors from "cors";
 import { createHash, randomUUID } from "crypto";
 import { eq } from "drizzle-orm";
+import { toMilliseconds } from "duration-fns";
 import { Router } from "express";
 import pRetry from "p-retry";
 import { findInferenceNode } from "./_common.js";
@@ -87,11 +88,15 @@ export default Router()
     const gptSessionId = randomUUID();
     const result = await pRetry(
       async () => {
-        for await (const chunk of inferenceNodeApi.create(node.baseUrl, {
-          id: gptSessionId,
-          initialPrompt: body.output.initialPrompt,
-          dumpSession,
-        })) {
+        for await (const chunk of inferenceNodeApi.create(
+          node.baseUrl,
+          {
+            id: gptSessionId,
+            initialPrompt: body.output.initialPrompt,
+            dumpSession,
+          },
+          { abortSignal: timeoutSignal(toMilliseconds({ minutes: 2 })) },
+        )) {
           switch (chunk.type) {
             case "Decode":
               res.write(