+++ packages/client/src/lib/tauri/gpt/infer.ts
@@ -1,7 +1,7 @@
 import { InferenceOptionsSchema } from "@/lib/ai/common";
 import { v } from "@/lib/valibot";
 import { invoke } from "@tauri-apps/api";
-import { listen } from "@tauri-apps/api/event";
+import { emit, listen } from "@tauri-apps/api/event";
 
 type Response = string;
 
@@ -16,6 +16,7 @@ type InferenceContent = {
 const COMMAND_NAME = "gpt_infer";
 const INFERENCE_EVENT_NAME = "app://gpt/inference";
 const DECODING_EVENT_NAME = "app://gpt/decoding";
+const ABORT_INFERENCE_EVENT_NAME = "app://gpt/abort-inference";
 
 /**
  * Predict text. Does not update the KV cache.
@@ -27,6 +28,7 @@ export async function infer(
   inferOptions: v.InferInput<typeof InferenceOptionsSchema> = {},
   decodeCallback?: (event: DecodingProgress) => void,
   inferenceCallback?: (event: InferenceContent) => void,
+  abortSignal?: AbortSignal,
 ): Promise<Response> {
   const unlistenDecode = decodeCallback
     ? await listen(DECODING_EVENT_NAME, (event) => {
@@ -39,6 +41,15 @@ export async function infer(
         inferenceCallback(event.payload as InferenceContent);
       })
     : undefined;
+
+  if (abortSignal) {
+    abortSignal.addEventListener("abort", () => {
+      console.log("aborting inference");
+      emit(ABORT_INFERENCE_EVENT_NAME, gptId).then(() => {
+        console.log(`Sent ${ABORT_INFERENCE_EVENT_NAME} event to ${gptId}.`);
+      });
+    });
+  }
 
   const result = (await invoke(COMMAND_NAME, {
     gptId,